requirements.txt
# GridOKB-USDT依赖库 (适配Python 3.13.1)
aiohttp>=3.9.1
aiofiles>=23.2.1
python-okx>=0.3.8
numpy>=1.26.0
python-dotenv>=1.0.0
uvicorn>=0.25.0 # ASGI server, potentially needed for aiohttp
requests>=2.31.0
tenacity>=8.2.3
psutil>=5.9.6
websockets>=12.0.0
pydantic>=2.5.0
pandas>=2.2.0
jinja2>=3.1.2
loguru>=0.7.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6 